Make call to VerifyTypes explicit
