+++ src/Compilers/CSharp/Test/Symbol/DeclarationTests.cs
@@ -288,7 +288,7 @@ public class B
 
             var countedTree = new CountedSyntaxTree(foreignType);
 
-            var compilation = CreateStandardCompilation(new SyntaxTree[] { underlyingTree, countedTree }, skipVerify: true);
+            var compilation = CreateStandardCompilation(new SyntaxTree[] { underlyingTree, countedTree });
 
             var type = compilation.Assembly.GlobalNamespace.GetTypeMembers().First();
             Assert.Equal(1, countedTree.AccessCount);   // parse once to build the decl table
